Shift emphasis from title to work
So much of the usefulness of Git & GitHub for library work extends to any kind of role in the library. By decreasing the emphasis on the professional title "librarian" and increasing focus on the work that you can accomplish with these tools, this lesson is friendlier to any participants, not just those with an MLS degree or who have "professional" positions.