Remove jargon
I suggest describing a Git repository as a "file" rather than a "data structure" (line 30, first sentence under "Creating a repository"), as this is a simpler concept for most users. This came up as feedback when I gave a practice version of this class as part of earning my Carpentries instructor certification.