Remove Notepad for Windows
Hi, just taught this today and wondered why Windows is isolated with notepad, since nano was used in the Unix Shell lesson.
So I pulled the line mentioning notepad, edited the statement to be Win, Appl and LInux, and removed the box of code.

It's a bit odd because we don't return to a text editor through the command line.  I like the way the learners can see the same effect from the gui interface of editing a file to push and commit.  So maybe there's another line left over in a different part of the lesson to also update about notepad and text editing.